Refactor Parser into Parser+ParserDaemon
