+++ habitat/tests/test_parser/test_parser.py
@@ -28,8 +28,6 @@ import M2Crypto
 from copy import deepcopy
 from nose.tools import assert_raises
 
-from ...utils import immortal_changes
-
 from ... import parser
 
 
@@ -50,13 +48,11 @@ class TestParser(object):
             "couch_uri": "http://localhost:5984", "couch_db": "test"}
 
         self.m.StubOutWithMock(parser, 'couchdbkit')
-        self.m.StubOutWithMock(parser, 'immortal_changes')
         self.mock_server = self.m.CreateMock(couchdbkit.Server)
         self.mock_db = self.m.CreateMock(couchdbkit.Database)
         parser.couchdbkit.Server("http://localhost:5984")\
                 .AndReturn(self.mock_server)
         self.mock_server.__getitem__("test").AndReturn(self.mock_db)
-        self.mock_db.info().AndReturn({"update_seq": 191238})
 
         self.m.ReplayAll()
         self.parser = parser.Parser(self.parser_config)
@@ -135,79 +131,6 @@ class TestParser(object):
         # This actually tested by setup(), since the parser needs to be
         # initialised with CouchDB mocks in all other tests.
         assert self.parser.db == self.mock_db
-
-    def test_run_calls_wait_and_uses_update_seq(self):
-        c = self.m.CreateMock(immortal_changes.Consumer)
-        parser.immortal_changes.Consumer(self.parser.db).AndReturn(c)
-        c.wait(self.parser._couch_callback, filter="habitat/unparsed",
-               since=191238, include_docs=True, heartbeat=1000)
-        self.m.ReplayAll()
-        self.parser.run()
-        self.m.VerifyAll()
-
-    def test_couch_callback(self):
-        result = {'doc': {'hello': 'world'}, 'seq': 1}
-        parsed = {'hello': 'parser'}
-        self.m.StubOutWithMock(self.parser, 'parse')
-        self.m.StubOutWithMock(self.parser, '_save_updated_doc')
-        self.parser.parse(result['doc']).AndReturn(parsed)
-        self.parser._save_updated_doc(parsed)
-        self.m.ReplayAll()
-        self.parser._couch_callback(result)
-        self.m.VerifyAll()
-
-    def test_saving_saves(self):
-        orig_doc = {"_id": "id", "receivers": [1], 'data': {'a': 1}}
-        parsed_doc = deepcopy(orig_doc)
-        parsed_doc['data']['b'] = 2
-        self.parser.db.__getitem__('id').AndReturn(orig_doc)
-        self.parser.db.save_doc(parsed_doc)
-        self.m.ReplayAll()
-        self.parser._save_updated_doc(parsed_doc)
-        self.m.VerifyAll()
-
-    def test_saving_merges(self):
-        orig_doc = {"_id": "id", "receivers": [1], 'data': {'a': 1}}
-        parsed_doc = deepcopy(orig_doc)
-        parsed_doc['data']['b'] = 2
-        updated_doc = deepcopy(orig_doc)
-        updated_doc['receivers'].append(2)
-        merged_doc = deepcopy(parsed_doc)
-        merged_doc['receivers'] = deepcopy(updated_doc['receivers'])
-        self.parser.db.__getitem__('id').AndReturn(updated_doc)
-        self.parser.db.save_doc(merged_doc)
-        self.m.ReplayAll()
-        self.parser._save_updated_doc(parsed_doc)
-        self.m.VerifyAll()
-
-    def test_saving_merges_after_conflict(self):
-        orig_doc = {"_id": "id", "receivers": [1], 'data': {'a': 1}}
-        parsed_doc = deepcopy(orig_doc)
-        parsed_doc['data']['b'] = 2
-        updated_doc = deepcopy(orig_doc)
-        updated_doc['receivers'].append(2)
-        merged_doc = deepcopy(parsed_doc)
-        merged_doc['receivers'] = deepcopy(updated_doc['receivers'])
-        self.parser.db.__getitem__('id').AndReturn(orig_doc)
-        self.parser.db.save_doc(parsed_doc).AndRaise(
-            couchdbkit.exceptions.ResourceConflict())
-        self.parser.db.__getitem__('id').AndReturn(updated_doc)
-        self.parser.db.save_doc(merged_doc)
-        self.m.ReplayAll()
-        self.parser._save_updated_doc(parsed_doc)
-        self.m.VerifyAll()
-
-    def test_saving_quits_after_many_conflicts(self):
-        orig_doc = {"_id": "id", "receivers": [1], 'data': {'a': 1}}
-        parsed_doc = deepcopy(orig_doc)
-        parsed_doc['data']['b'] = 2
-        for i in xrange(30):
-            self.parser.db.__getitem__('id').AndReturn(orig_doc)
-            self.parser.db.save_doc(parsed_doc).AndRaise(
-                couchdbkit.exceptions.ResourceConflict())
-        self.m.ReplayAll()
-        assert_raises(RuntimeError, self.parser._save_updated_doc, parsed_doc)
-        self.m.VerifyAll()
 
     def test_looks_for_config_doc(self):
         callsign = "habitat"
